PageForm: extract PageTitleField
