+++ front/app/modules/commercial/customizable_navbar/admin/components/PoliciesSubtitle.tsx
@@ -7,12 +7,15 @@ import { StyledLink } from 'containers/Admin/settings/policies';
 import { FormattedMessage } from 'utils/cl-intl';
 import messages from './messages';
 
+// utils
+import { NAVIGATION_PATH } from '../containers';
+
 export default () => (
   <FormattedMessage
     {...messages.policiesSubtitlePremium}
     values={{
       navigationLink: (
-        <StyledLink to="/admin/settings/navigation">
+        <StyledLink to={NAVIGATION_PATH}>
           <FormattedMessage {...messages.linkToNavigation} />
         </StyledLink>
       ),
